+++ mods/iteminteractions/1.20.4/Common/src/main/java/fuzs/iteminteractions/api/v1/ContainerItemHelper.java
@@ -3,26 +3,22 @@ package fuzs.iteminteractions.api.v1;
 import fuzs.iteminteractions.api.v1.provider.ItemContainerProvider;
 import fuzs.iteminteractions.impl.world.inventory.SimpleSlotContainer;
 import fuzs.iteminteractions.impl.world.item.container.ContainerItemHelperImpl;
+import fuzs.iteminteractions.api.v1.provider.ItemContainerBehavior;
 import fuzs.iteminteractions.impl.world.item.container.ItemInteractionHelper;
-import net.minecraft.core.NonNullList;
 import net.minecraft.world.SimpleContainer;
 import net.minecraft.world.item.DyeColor;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.Objects;
 import java.util.function.IntFunction;
 
 public interface ContainerItemHelper {
     ContainerItemHelper INSTANCE = new ContainerItemHelperImpl();
 
-    default ItemContainerProvider getItemContainerProviderOrThrow(ItemStack stack) {
-        ItemContainerProvider provider = this.getItemContainerProvider(stack);
-        Objects.requireNonNull(provider, "provider is null");
-        return provider;
-    }
+    ItemContainerBehavior getItemContainerBehavior(ItemStack itemStack);
 
-    @Nullable ItemContainerProvider getItemContainerProvider(ItemStack stack);
+    ItemContainerBehavior getItemContainerBehavior(Item item);
 
     default SimpleContainer loadItemContainer(ItemStack stack, ItemContainerProvider provider, int inventorySize, boolean allowSaving) {
         return this.loadItemContainer(stack, provider, inventorySize, allowSaving, ItemInteractionHelper.TAG_ITEMS);
@@ -33,10 +29,6 @@ public interface ContainerItemHelper {
     }
 
     SimpleContainer loadItemContainer(ItemStack stack, ItemContainerProvider provider, IntFunction<SimpleContainer> containerFactory, boolean allowSaving, String nbtKey);
-
-    int getItemWeight(ItemStack stack);
-
-    NonNullList<ItemStack> getListFromContainer(SimpleContainer container);
 
     float[] getBackgroundColor(@Nullable DyeColor backgroundColor);
 }