+++ mods/iteminteractions/1.20.4/Common/src/main/java/fuzs/iteminteractions/impl/client/helper/ItemDecorationHelper.java
@@ -3,9 +3,9 @@ package fuzs.iteminteractions.impl.client.helper;
 import com.google.common.collect.Maps;
 import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.blaze3d.vertex.Tesselator;
-import fuzs.iteminteractions.api.v1.provider.ItemContainerProvider;
 import fuzs.iteminteractions.impl.ItemInteractions;
 import fuzs.iteminteractions.impl.config.ClientConfig;
+import fuzs.iteminteractions.api.v1.provider.ItemContainerBehavior;
 import fuzs.iteminteractions.impl.world.item.container.ItemContainerProviders;
 import fuzs.puzzleslib.api.client.init.v1.DynamicItemDecorator;
 import fuzs.puzzleslib.api.core.v1.ModLoaderEnvironment;
@@ -24,7 +24,7 @@ import java.util.Map;
 import java.util.function.BooleanSupplier;
 
 public class ItemDecorationHelper {
-    private static final Map<ItemContainerProvider, DynamicItemDecorator> DECORATORS_CACHE = Maps.newIdentityHashMap();
+    private static final Map<ItemContainerBehavior, DynamicItemDecorator> DECORATORS_CACHE = Maps.newIdentityHashMap();
 
     @Nullable
     private static Slot activeSlot;
@@ -67,11 +67,11 @@ public class ItemDecorationHelper {
     }
 
     public static void render(GuiGraphics guiGraphics, Font font, ItemStack stack, int itemPosX, int itemPosY) {
-        ItemContainerProvider provider = ItemContainerProviders.INSTANCE.get(stack);
-        if (provider != null) {
+        ItemContainerBehavior behavior = ItemContainerProviders.INSTANCE.get(stack);
+        if (!behavior.isEmpty()) {
             resetRenderState();
-            DynamicItemDecorator itemDecorator = DECORATORS_CACHE.computeIfAbsent(provider, $ -> ItemDecorationHelper.getDynamicItemDecorator((AbstractContainerScreen<?> screen, ItemStack containerStack, ItemStack carriedStack) -> {
-                return ItemDecoratorType.getItemDecoratorType(provider, containerStack, carriedStack, screen.minecraft.player);
+            DynamicItemDecorator itemDecorator = DECORATORS_CACHE.computeIfAbsent(behavior, $ -> ItemDecorationHelper.getDynamicItemDecorator((AbstractContainerScreen<?> screen, ItemStack containerStack, ItemStack carriedStack) -> {
+                return ItemDecoratorType.getItemDecoratorType(behavior, containerStack, carriedStack, screen.minecraft.player);
             }, () -> ItemInteractions.CONFIG.get(ClientConfig.class).containerItemIndicator));
             if (itemDecorator.renderItemDecorations(guiGraphics, font, stack, itemPosX, itemPosY)) {
                 resetRenderState();