+++ mods/iteminteractions/1.20.4/Common/src/main/java/fuzs/iteminteractions/mixin/ItemStackMixin.java
@@ -1,6 +1,6 @@
 package fuzs.iteminteractions.mixin;
 
-import fuzs.iteminteractions.api.v1.provider.ItemContainerProvider;
+import fuzs.iteminteractions.api.v1.provider.ItemContainerBehavior;
 import fuzs.iteminteractions.impl.world.item.container.ItemContainerProviders;
 import fuzs.iteminteractions.impl.world.item.container.ItemInteractionHelper;
 import fuzs.puzzleslib.api.core.v1.Proxy;
@@ -23,10 +23,18 @@ abstract class ItemStackMixin {
     @Inject(method = "overrideStackedOnOther", at = @At("HEAD"), cancellable = true)
     public void overrideStackedOnOther(Slot slot, ClickAction clickAction, Player player, CallbackInfoReturnable<Boolean> callback) {
         ItemStack containerStack = ItemStack.class.cast(this);
-        ItemContainerProvider provider = ItemContainerProviders.INSTANCE.get(containerStack);
-        if (provider != null && provider.allowsPlayerInteractions(containerStack, player)) {
-            boolean result = ItemInteractionHelper.overrideStackedOnOther(() -> provider.getItemContainer(containerStack, player, true), slot, clickAction, player, stack -> provider.getAcceptableItemCount(containerStack, stack, player), provider::getMaxStackSize);
-            if (result) provider.broadcastContainerChanges(player);
+        ItemContainerBehavior behavior = ItemContainerProviders.INSTANCE.get(containerStack);
+        if (behavior.allowsPlayerInteractions(containerStack, player)) {
+            boolean result = ItemInteractionHelper.overrideStackedOnOther(() -> behavior.getItemContainer(containerStack,
+                            player
+                    ),
+                    slot,
+                    clickAction,
+                    player,
+                    stack -> behavior.getAcceptableItemCount(containerStack, stack, player),
+                    behavior.provider()::getMaxStackSize
+            );
+            if (result) behavior.provider().broadcastContainerChanges(player);
             callback.setReturnValue(result);
         }
     }
@@ -34,10 +42,21 @@ abstract class ItemStackMixin {
     @Inject(method = "overrideOtherStackedOnMe", at = @At("HEAD"), cancellable = true)
     public void overrideOtherStackedOnMe(ItemStack stackOnMe, Slot slot, ClickAction clickAction, Player player, SlotAccess slotAccess, CallbackInfoReturnable<Boolean> callback) {
         ItemStack containerStack = ItemStack.class.cast(this);
-        ItemContainerProvider provider = ItemContainerProviders.INSTANCE.get(containerStack);
-        if (provider != null && provider.allowsPlayerInteractions(containerStack, player)) {
-            boolean result = ItemInteractionHelper.overrideOtherStackedOnMe(() -> provider.getItemContainer(containerStack, player, true), stackOnMe, slot, clickAction, player, slotAccess, stack -> provider.getAcceptableItemCount(containerStack, stack, player), provider::getMaxStackSize);
-            if (result) provider.broadcastContainerChanges(player);
+        ItemContainerBehavior behavior = ItemContainerProviders.INSTANCE.get(containerStack);
+        if (behavior.allowsPlayerInteractions(containerStack, player)) {
+            boolean result = ItemInteractionHelper.overrideOtherStackedOnMe(() -> behavior.getItemContainer(
+                            containerStack,
+                            player
+                    ),
+                    stackOnMe,
+                    slot,
+                    clickAction,
+                    player,
+                    slotAccess,
+                    stack -> behavior.getAcceptableItemCount(containerStack, stack, player),
+                    behavior.provider()::getMaxStackSize
+            );
+            if (result) behavior.provider().broadcastContainerChanges(player);
             callback.setReturnValue(result);
         }
     }
@@ -45,9 +64,9 @@ abstract class ItemStackMixin {
     @Inject(method = "getTooltipImage", at = @At("HEAD"), cancellable = true)
     public void getTooltipImage(CallbackInfoReturnable<Optional<TooltipComponent>> callback) {
         ItemStack containerStack = ItemStack.class.cast(this);
-        ItemContainerProvider provider = ItemContainerProviders.INSTANCE.get(containerStack);
-        if (provider != null && provider.canProvideTooltipImage(containerStack, Proxy.INSTANCE.getClientPlayer())) {
-            callback.setReturnValue(provider.getTooltipImage(containerStack, Proxy.INSTANCE.getClientPlayer()));
+        ItemContainerBehavior behavior = ItemContainerProviders.INSTANCE.get(containerStack);
+        if (behavior.canProvideTooltipImage(containerStack, Proxy.INSTANCE.getClientPlayer())) {
+            callback.setReturnValue(behavior.getTooltipImage(containerStack, Proxy.INSTANCE.getClientPlayer()));
         }
     }
 }